+++ custom_components/unifiprotect/entity.py
@@ -1,19 +1,38 @@
 """Shared Entity definition for Unifi Protect Integration."""
 from __future__ import annotations
 
-from typing import Any
+import collections
+from datetime import timedelta
+import hashlib
+import logging
+from random import SystemRandom
+from typing import Any, Final
 
 from homeassistant.const import ATTR_ATTRIBUTION
 from homeassistant.core import callback
 import homeassistant.helpers.device_registry as dr
+from homeassistant.helpers.dispatcher import (
+    async_dispatcher_connect,
+    async_dispatcher_send,
+)
 from homeassistant.helpers.entity import DeviceInfo, Entity, EntityDescription
 from pyunifiprotect import ProtectApiClient
 from pyunifiprotect.data.base import ProtectAdoptableDeviceModel, ProtectDeviceModel
 from pyunifiprotect.data.nvr import NVR
 from pyunifiprotect.data.types import ModelType
 
-from .const import ATTR_DEVICE_MODEL, DEFAULT_ATTRIBUTION, DEFAULT_BRAND, DOMAIN
+from .const import (
+    ATTR_DEVICE_MODEL,
+    DEFAULT_ATTRIBUTION,
+    DEFAULT_BRAND,
+    DOMAIN,
+    EVENT_UPDATE_TOKENS,
+)
 from .data import UnifiProtectData
+
+TOKEN_CHANGE_INTERVAL: Final = timedelta(minutes=1)
+_RND: Final = SystemRandom()
+_LOGGER = logging.getLogger(__name__)
 
 
 class UnifiProtectEntity(Entity):
@@ -73,7 +92,9 @@ class UnifiProtectEntity(Entity):
     @property
     def extra_state_attributes(self) -> dict[str, Any]:
         """Return common attributes"""
+        attrs = super().extra_state_attributes or {}
         return {
+            **attrs,
             ATTR_ATTRIBUTION: DEFAULT_ATTRIBUTION,
             ATTR_DEVICE_MODEL: self.device.type,
         }
@@ -99,8 +120,61 @@ class UnifiProtectEntity(Entity):
 
     async def async_added_to_hass(self) -> None:
         """When entity is added to hass."""
+        await super().async_added_to_hass()
         self.async_on_remove(
             self.protect_data.async_subscribe_device_id(
                 self.device.id, self._async_updated_event
+            )
+        )
+
+
+class AccessTokenMixin(Entity):
+    """Adds access_token attribute and provides access tokens for use for anonymous views."""
+
+    @property
+    def access_tokens(self) -> collections.deque[str]:
+        """Get valid access_tokens for current entity"""
+        assert isinstance(self, UnifiProtectEntity)
+        return self.protect_data.async_get_or_create_access_tokens(self.entity_id)
+
+    @callback
+    def _async_update_and_write_token(self):
+        _LOGGER.debug("Updating access tokens for %s", self.entity_id)
+        self.async_update_token()
+        self.async_write_ha_state()
+
+    @callback
+    def async_update_token(self) -> None:
+        """Update the used token."""
+        self.access_tokens.append(
+            hashlib.sha256(_RND.getrandbits(256).to_bytes(32, "little")).hexdigest()
+        )
+
+    @callback
+    def _trigger_update_tokens(self, *args, **kwargs):
+        assert isinstance(self, UnifiProtectEntity)
+        async_dispatcher_send(
+            self.hass,
+            f"{EVENT_UPDATE_TOKENS}-{self.entity_description.key}-{self.entity_id}",
+        )
+
+    async def async_added_to_hass(self) -> None:
+        """Run when entity about to be added to hass.
+
+        Injects callbacks to update access tokens automatically
+        """
+        await super().async_added_to_hass()
+
+        self.async_update_token()
+        self.async_on_remove(
+            async_dispatcher_connect(
+                self.hass,
+                f"{EVENT_UPDATE_TOKENS}-{self.entity_id}",
+                self._async_update_and_write_token,
+            )
+        )
+        self.async_on_remove(
+            self.hass.helpers.event.async_track_time_interval(
+                self._trigger_update_tokens, TOKEN_CHANGE_INTERVAL
             )
         )